+++ code/algorithms/rttbMaxOHxVolumeToDoseMeasureCollectionCalculator.cpp
@@ -37,7 +37,7 @@ namespace rttb
 			DoseTypeGy resultDose = 0;
 
 			double countVoxels = 0;
-			for (auto i = _doseVector.size() - 1; i != -1; i--)
+			for (auto i = _doseVector.size() - 1; i!=static_cast<size_t>(-1); i--)
 			{
 				countVoxels += _voxelProportionVector.at(i);
